requirements.txt
loguru==0.7.0
openai==0.27.9
python-telegram-bot==20.4
